% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{constrain}
\alias{constrain}
\alias{co}
\title{Specify constraints on data sets or observations}
\usage{
constrain(
  expr,
  name = to_character(expr),
  vars = detect_vars(expr),
  target = detect_target(expr),
  type = c(NA, "context", "population", "admissibility")
)

co(
  expr,
  name = to_character(expr),
  vars = detect_vars(expr),
  target = detect_target(expr),
  type = c(NA, "context", "population", "admissibility")
)
}
\arguments{
\item{expr}{(character | formula) \cr ...}

\item{name}{(character) \cr ...}

\item{vars}{(character) \cr ...}

\item{target}{(character) \cr ...}

\item{type}{(character) \cr ...}
}
\value{
(mldesign_constraint) \cr ...
}
\description{
A short description...
}
\examples{
constrain("test$age > 18", type = "population")
constrain(~ (test$year - train$year) \%>\% dplyr::between(1, 3), type = "context")
constrain("nrow(train) > 100")
}
