% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_splits.R
\name{derive_splits}
\alias{derive_splits}
\title{Derive a data split for a given estimand.}
\usage{
derive_splits(spec, data)
}
\arguments{
\item{spec}{(mldesign_spec) \cr a \code{mldesign_spec} object, i.e. a
\code{mldesign_estimand} object created via specify \code{\link{specify_estimand}} or a
\code{mldesign_method} object created via specify \code{\link{specify_method}}}

\item{data}{(data.frame) \cr the complete data set.}
}
\value{
(mldesign_splits)
}
\description{
Derive a data split for a given estimand.
}
