% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_method.R
\name{specify_method}
\alias{specify_method}
\title{Specify a data splitting method}
\usage{
specify_method(method, ..., default = FALSE, name = method)
}
\arguments{
\item{method}{(character) \cr}

\item{...}{(any) \cr named arguments to function name}

\item{default}{(logical) \cr Should default arguments be used? (default: FALSE)}

\item{name}{(character) \cr name of the data splitting method, equal to \code{method} by default.}
}
\value{
an object of class \code{mldesign_method}
}
\description{
Specify a data splitting method
}
\examples{
\dontrun{
specify_method()
}
specify_method("cv")
specify_method("cv", default = TRUE)
specify_method("cv", n_folds = 10)
}
