% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_nested.R
\name{specify_nested}
\alias{specify_nested}
\alias{nest}
\title{Specify a nested design}
\usage{
specify_nested(outer, inner = outer, name = NULL)

nest(outer, inner = outer, name = NULL)
}
\arguments{
\item{outer}{(mldesign_estimand | mldesign_method) \cr outer design specification}

\item{inner}{(mldesign_estimand | mldesign_method) \cr inner design specification}

\item{name}{(character) \cr an optional name for the design}
}
\value{
(mldesign_splits)
}
\description{
Specify a nested design
}
\examples{
nest(
specify_estimand(constrain(~train$country != test$country)),
specify_method("cv", n_folds=5)
)
}
